// src/app/api_codegen/model/updateZoneDto.ts
/**
 * Invicto API
 * REST API Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface UpdateZoneDto { 
    id: number;
    name?: string;
    description?: string;
    doorIds?: Array<number>;
    zoneIds?: Array<number>;
}